Adds /next route as well as buttons that use it
The next route redirects to either the first ticket assigned to a user
or the first unassigned ticket if none are assigned.

The Resolve and Next button will appear on ticket pages you're assigned to,
which will resolve the current ticket and then redirect to next.

The Next Ticket button will appear on ticket pages that are queued or resolved
or are assigned to a staff member other than you.

The regular Resolve button now redirects back to the queue after resolving.